+++ inst/shiny-apps/birdrecoveries/ui.R
@@ -1,32 +1,27 @@
+library(leaflet)
+library(DT)
+#library(shinydashboard)
+
 shinyUI(fluidPage(
   theme = shinythemes::shinytheme("spacelab"),
   tags$head(tags$link(rel="shortcut icon", href="favicon.ico")),
-  #  titlePanel("Fågeltrender - Totaler"),
+  #titlePanel("Swedish Bird Recoveries / Återfynd"),
+  tags$style(type = "text/css", "#map {height: calc(100vh - 80px) !important;}"),
   sidebarLayout(
     sidebarPanel(
-      img(src = "logo.png", width = 50),
+    	img(src = "nrm-logo.png", width = 50),
+    	img(src = "logo.png", width = 50),
       hr(),
-      selectizeInput("name", label = "Urval av fågelarter:",
-       choices = unique(birdrecoveries$name),
-       multiple = TRUE,
-       options = list(maxItems = 20, placeholder = "Välj namn")#,
-      )
-
+      uiOutput("species"),
+      uiOutput("country"),
+      uiOutput("lats"),
+      uiOutput("lons"),
+      uiOutput("source"),
+      uiOutput("lang")
     ),
+
     mainPanel(
-      tabsetPanel(
-        tabPanel("Tabell",
-          helpText("Nuvarande urval:"),
-          br(),
-          DT::dataTableOutput("table")
-        ),
-        tabPanel("Källa",
-          helpText("Ladda ned nuvarande data i CSV:"),
-          fluidRow(p(class = "text-center",
-            downloadButton("dl", label = "Hämta all data"))
-          )
-        )
-      )
+    	uiOutput("mytabs")
     )
   )
 ))